String with numbers at its end
Given a string (S) that is appended with a number at last. You need to find whether the length of string excluding that number is equal to that number. For example for “helloworld10”, answer is True as helloworld consist of 10 letters. 
Examples:

Input:  S = "geeks5"
Output:  1
Explanation : As geeks is of 5 length and at 
              last number is also 5.

Input:  S = "geeksforgeeks15"
Output:  0
Explanation: As geeksforgeeks is of 13 length and
             at last number is 15 i.e. not equal