Gemstones
### [Challenge Name: Gemstones](/challenges/gem-stones)


John has discovered various rocks. Each rock is composed of various elements, and each element is represented by a lower-case Latin letter from 'a' to 'z'. An element can be present multiple times in a rock. An element is called a _gem-element_ if it occurs at least once in each of the rocks.

Given the list of $N$ rocks with their compositions, display the number of gem-elements that exist in those rocks.