Alternating Characters
### [Challenge Name: Alternating Characters ](/challenges/alternating-characters)


You are given a string containing characters $A$ and $B$ only, your task is to change it into a string such that every two consecutive characters are different. To do this, you are allowed to delete one or more characters in the string.  

Your task is to find the minimum number of required deletions.

For example, string $ABAA$ should be changed to $ABA$ by deleting one character $A$.