Anagram
### [Challenge Name: Anagram](/challenges/anagram)


Sid is obsessed with reading short stories. Being a CS student, he is doing some interesting frequency analysis with the books. He chooses strings $S1$ and $S2$ in such a way that $|len(S1) -len(S2)| \le 1$.  

Your task is to help him find the minimum number of characters of the first string he needs to change to enable him to make it an [anagram](https://en.wikipedia.org/wiki/Anagram) of the second string.  

*Note:* A word _x_ is an anagram of another word _y_ if we can produce _y_ by rearranging the letters of _x_.